Encode overloading and rewrite inheritance support.
