+++ examples/known-problems/FunctionProblem.pvl
@@ -1,0 +1,26 @@
+// -*- tab-width:2 ; indent-tabs-mode:nil -*-
+
+//:: cases FunctionProblemFull
+//:: suite problem-fail
+//:: tools silicon
+//:: verdict Pass
+
+
+/**
+ * This file triggers the Silver warning:
+ *  abstract predicates cannot be (un)folded
+ * this error is however not reported by VerCors.
+ */
+
+class problem {
+
+  resource p();
+  
+  requires p();
+  int f()=unfolding p() in 0;
+  
+  requires p();
+  void m(){
+    unfold p();
+  }
+}
