+++ examples/threads/VerifiedMain-E2.java
@@ -1,5 +1,6 @@
 // -*- tab-width:2 ; indent-tabs-mode:nil -*-
 //:: cases IncrThread-E2
+//:: suite skip-travis
 //:: tool silicon
 //:: verdict Fail
