+++ examples/witnesses/ListAppendASyncDef.java
@@ -12,7 +12,7 @@
   predicate.
 */
 
-class List {
+final class List {
 
   public int val;
   public List next;
