+++ vercors/src/main/java/vct/col/util/JavaTypeCheck.java
@@ -1,0 +1,18 @@
+package vct.col.util;
+
+import vct.col.ast.ProgramUnit;
+
+/**
+ * This class implements type checking of object oriented programs
+ * that may use inheritance and/or overloading.
+ * 
+ * @author Stefan Blom
+ *
+ */
+public class JavaTypeCheck extends AbstractTypeCheck {
+
+  public JavaTypeCheck(ProgramUnit arg) {
+    super(arg);
+  }
+
+}
